Update and take advantage of our ability to override RequireJS paths.
We had a mechanism to, at runtime, generate a RequireJS config that
would override the base paths and, instead, use the hashed versions of
assets for things that RequireJs was loading on demand.

We've now moved that out of the coureware.html base template and into
main.html so that more pages actually benefit from it.

As well, we've added some of the heavy hitters for unhashed assets,
namely moment.min.js, to these overrides which should allow better
caching for end users.  We'll be able to add more things to the override
list in the future as they crop up.